Reformat code (sheesh Hyper all this stuff is new)

+++ src/main/java/com/hyperdash/firmaciv/entity/custom/VehicleHelperEntities/AbstractCompartmentEntity.java
@@ -132,7 +132,7 @@ public class AbstractCompartmentEntity extends Entity {
                     this.move(MoverType.SELF, this.getDeltaMovement());
                     float f1 = 0.98F;
 
-                    this.setDeltaMovement(this.getDeltaMovement().multiply((double) f1, 0.98D, (double) f1));
+                    this.setDeltaMovement(this.getDeltaMovement().multiply(f1, 0.98D, f1));
                     if (this.onGround()) {
                         Vec3 vec31 = this.getDeltaMovement();
                         if (vec31.y < 0.0D) {
